+++ bench/whereblocks.py
@@ -4,6 +4,7 @@ import numexpr as ne
 import bcolz
 import time
 import cProfile
+import inspect
 
 
 bcolz.defaults.cparams['shuffle'] = bcolz.SHUFFLE
@@ -13,9 +14,10 @@ bcolz.defaults.cparams['cname'] = 'lz4'
 bcolz.defaults.cparams['clevel'] = 5
 
 N = 1e8
-a = np.arange(N)
-b = np.arange(N)
-ct = bcolz.ctable([a,b], names=["a", "b"])
+LMAX = 1e3
+a1 = np.arange(N)
+b1 = np.arange(N)
+ct = bcolz.ctable([a1,b1], names=["a", "b"])
 
 
 def do_cprofile(func):
@@ -27,7 +29,7 @@ def do_cprofile(func):
             profile.disable()
             return result
         finally:
-            profile.print_stats()
+            profile.print_stats(sort='cumulative')
     return profiled_func
 
 
@@ -44,33 +46,38 @@ def timefunc(f):
 
 @timefunc
 def where0():
-    return sum(a[i] for i in np.where((a > 5) & (b < 1e6))[0])
+    return sum(a1[i] for i in np.where((a1 > 5) & (b1 < LMAX))[0])
 
 @timefunc
 #@do_cprofile
 def where1():
-    #return sum(r[0] for r in ct.where("(a > 5) & (b < 10)"))
-    return sum(r[0] for r in ct.where("(a > 5) & (b < 1e6)", out_flavor=tuple))
+    return sum(r[0] for r in ct.where("(a > 5) & (b < LMAX)",
+                                      out_flavor=tuple))
+@timefunc
+#@do_cprofile
+def where2():
+    return sum(r[0] for r in ct.where("(a1 > 5) & (b1 < LMAX)",
+                                      out_flavor=tuple))
 
 @timefunc
 #@do_cprofile
 def whereblocks():
     sum = 0.
-    for r in ct.whereblocks("(a > 5) & (b < 1e6)", blen=None):
-    #for r in ct.whereblocks("(a > 5) & (b < 1e6)", blen=ct['a'].chunklen*10):
-    #for r in ct.whereblocks("(a > 5) & (b < 1e6)", blen=1000):
+    for r in ct.whereblocks("(a > 5) & (b < LMAX)", blen=None):
+    #for r in ct.whereblocks("(a > 5) & (b < LMAX)", blen=ct['a'].chunklen*10):
+    #for r in ct.whereblocks("(a > 5) & (b < LMAX)", blen=1000):
         sum += r['a'].sum()
     return sum
 
 @timefunc
 #@do_cprofile
 def fetchwhere_bcolz():
-    return ct.fetchwhere("(a > 5) & (b < 1e6)", out_flavor='bcolz')['a'].sum()
+    return ct.fetchwhere("(a > 5) & (b < LMAX)", out_flavor='bcolz')['a'].sum()
 
 @timefunc
 #@do_cprofile
 def fetchwhere_numpy():
-    return ct.fetchwhere("(a > 5) & (b < 1e6)", out_flavor='numpy')['a'].sum()
+    return ct.fetchwhere("(a > 5) & (b < LMAX)", out_flavor='numpy')['a'].sum()
 
 
 print repr(ct)
@@ -79,6 +86,9 @@ a0 = where0()
 print "a0:", a0
 a1 = where1()
 assert a0 == a1
+# a1 = where2()
+# print "a1:", a1
+# assert a0 == a1
 a1 = whereblocks()
 assert a0 == a1
 a1 = fetchwhere_bcolz()